+++ steel/src/primitives/ports.rs
@@ -1,8 +1,8 @@
 use crate::gc::Gc;
-use crate::port::SteelPort;
 use crate::rerrs::{ErrorKind, SteelErr};
 use crate::rvals::{Result, SteelVal};
 use crate::stop;
+use crate::values::port::SteelPort;
 
 pub struct PortOperations {}
 impl PortOperations {
